Create build.bat (local build script)
This batch script helps to build SGDK if PATH environmental variables are not setup for GDK already.                                    

(1) When run the GDK and GDK_WIN EnvVars will be configured temporarily for the local directory.                              
(2) The User will be asked whether to temporarily empty PATH. If yes, PATH will be cleared for the the local script's execution only.   
(3) SGDK clean script will be launched and then SGDK will be rebuild.