Improve assembly using macros and preprocessor
This change will create functions in their own .text section so that
the linker can garbage collect them (via --gc-sections). This reduces
ROM usage by a few KBs in case the functions are not used.

TODO: Do the same for the other .s files under /src/, for now just did
the big ones to save space