Clean-up Dockerfile
This re-works the Dockerfile to allow for a smaller size (saves about 20 MB) as
ewll as adds ARGs so that new version of SGDK_wine and the JDK version can be
tested before permanent changes are made to the Dockerfile. This also now
follows some Dockerfile best practices, mainly around minimizing of the number
of layers used and apt-get usage to minimize contianer size see
https://docs.docker.com/develop/develop-images/dockerfile_best-practices for
details.